refactor(mutation): extract workspace utilities into shared module
Move project-copying logic from mutation/runner.rs into a new
crates/forge/src/workspace.rs module for reuse by other features
that need isolated temp workspaces (e.g., brutalization).

Extracted functions: copy_project, copy_dir_recursive, symlink_dir,
symlink_nested_libs, relative_to_root, ensure_safe_relative_path.

Tests moved from mutation/runner.rs to workspace.rs.

Amp-Thread-ID: https://ampcode.com/threads/T-019c4d7b-acac-75ef-9f69-f3619f6941b6
